+++ library/SpatialDomains/Geometry.h
@@ -151,7 +151,7 @@ public:
         const Array<OneD, const NekDouble> &gloCoord,
         Array<OneD, NekDouble> &locCoord,
         NekDouble tol,
-        NekDouble &resid);
+        NekDouble &dist);
     SPATIAL_DOMAINS_EXPORT inline NekDouble GetLocCoords(
         const Array<OneD, const NekDouble> &coords,
         Array<OneD, NekDouble> &Lcoords);
@@ -224,7 +224,7 @@ protected:
     virtual bool v_ContainsPoint(const Array<OneD, const NekDouble> &gloCoord,
                                  Array<OneD, NekDouble> &locCoord,
                                  NekDouble tol,
-                                 NekDouble &resid);
+                                 NekDouble &dist);
 
     virtual NekDouble v_GetCoord(const int i,
                                  const Array<OneD, const NekDouble> &Lcoord);
@@ -458,8 +458,8 @@ inline bool Geometry::ContainsPoint(
     NekDouble tol)
 {
     Array<OneD,NekDouble> locCoord(GetCoordim(), 0.0);
-    NekDouble resid;
-    return v_ContainsPoint(gloCoord, locCoord, tol, resid);
+    NekDouble dist;
+    return v_ContainsPoint(gloCoord, locCoord, tol, dist);
 }
 
 /**
@@ -473,8 +473,8 @@ inline bool Geometry::ContainsPoint(
     Array<OneD, NekDouble> &locCoord,
     NekDouble tol)
 {
-    NekDouble resid;
-    return v_ContainsPoint(gloCoord, locCoord, tol, resid);
+    NekDouble dist;
+    return v_ContainsPoint(gloCoord, locCoord, tol, dist);
 }
 
 /**
@@ -497,7 +497,7 @@ inline bool Geometry::ContainsPoint(
  *                  that \f$\chi(\vec{\xi}) = \vec{x}\f$.
  * @param tol       The tolerance used to dictate the bounding box of the
  *                  standard coordinates \f$\vec{\xi}\f$.
- * @param resid     On exit, returns the minimum distance between @p gloCoord
+ * @param dist      On exit, returns the minimum distance between @p gloCoord
  *                  and the quadrature points inside the element.
  *
  * @return `true` if the coordinate @p gloCoord is contained in the element;
@@ -509,9 +509,9 @@ inline bool Geometry::ContainsPoint(
     const Array<OneD, const NekDouble> &gloCoord,
     Array<OneD, NekDouble> &locCoord,
     NekDouble tol,
-    NekDouble &resid)
+    NekDouble &dist)
 {
-    return v_ContainsPoint(gloCoord, locCoord, tol, resid);
+    return v_ContainsPoint(gloCoord, locCoord, tol, dist);
 }
 
 /**