+++ library/SpatialDomains/Geometry0D.cpp
@@ -59,9 +59,9 @@ int Geometry0D::v_GetShapeDim() const
 bool Geometry0D::v_ContainsPoint(const Array<OneD, const NekDouble> &gloCoord,
                                Array<OneD, NekDouble> &locCoord,
                                NekDouble tol,
-                               NekDouble &resid)
+                               NekDouble &dist)
 {
-    boost::ignore_unused(gloCoord, locCoord, tol, resid);
+    boost::ignore_unused(gloCoord, locCoord, tol, dist);
     NEKERROR(ErrorUtil::efatal,
              "This function has not been defined for 0D geometry");
     return false;